requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.11,<4.0
psycopg2>=2.9,<3.0
djangorestframework-simplejwt>=5.0,<6.0
